fix(Reporting): persist write-back migration + extract migration code
Fixes the legacy->canonical write-back, which silently failed for every
legacy ReportConfig: CouchDB rejects unknown underscore-prefixed
top-level members ("doc_validation: Bad special document member:
_legacyOriginal"), so the migrated doc was never persisted. Renamed the
rollback copy `_legacyOriginal` -> `legacyOriginal`. Mocked unit tests
missed this; added a regression assertion that the migrated body
introduces no illegal underscore-prefixed keys. Verified end-to-end:
legacy doc now migrates on first read, write-back persists, and is
idempotent (canonical docs are left untouched).

Also extracts all legacy-migration logic into a dedicated
ReportConfigLegacyMigration class (DefaultReportStorage delegates to it)
so it can be deleted wholesale once production docs are migrated, and
documents why normalizeLegacyArgs is still needed (it normalizes
per-calculation runtime args, independent of the config migration).
